requirements.txt
numpy >= 1.2
asana >= 1.0.0
pyperclip >= 1.8.2
python-dotenv >= 0.20.0
